+++ cmd/agt/main.go
@@ -90,6 +90,8 @@ func run(args []string, stdout, stderr io.Writer) int {
 		return cmdRedact(args[1:], stdout, stderr)
 	case "netguard":
 		return cmdNetguard(args[1:], stdout, stderr)
+	case "ratelimit":
+		return cmdRateLimit(args[1:], stdout, stderr)
 	case "doctor":
 		return cmdDoctor(args[1:], stdout, stderr)
 	case "quickstart":